mem-ruby: Optimal matrix iteration for topology weight tree
This change updates the way we iterate the weight tree when
calculating the shortest distance tree for a given topology.
Instead of parsing over all vnets for a given pair of source
and destination, we now split it into pairs for each vnet
reflective of how the algorithm accesses values.

Change-Id: Iad15c9b2406d2bf026b3ef2a70c0ba0fad4b03a5